+++ src/main/java/com/spring/localhongdae/restaurant/controller/RestaurantController.java
@@ -5,13 +5,12 @@ import com.spring.localhongdae.dto.PlaceDto;
 import com.spring.localhongdae.entity.City;
 import com.spring.localhongdae.entity.District;
 import com.spring.localhongdae.entity.Place;
-import com.spring.localhongdae.restaurant.service.RestaurantService;
+import com.spring.localhongdae.restaurant.service.RestaurantServiceImpl;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.repository.query.Param;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -26,7 +25,7 @@ import java.util.List;
 public class RestaurantController {
 
     @Autowired
-    RestaurantService restaurantService;
+    RestaurantServiceImpl restaurantService;
 
     @GetMapping("/restaurant/list")
     public ModelAndView getResturantList(HttpServletRequest request, @PageableDefault(page = 1, size = 5)Pageable pageable) {
